Move StringKeyedMap from util.go to dedicated concurret.go file
