+++ common/util.go
@@ -27,7 +27,6 @@ import (
 	"os"
 	"reflect"
 	"strings"
-	"sync"
 
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -475,36 +474,4 @@ func StructToMap(obj interface{}, output map[string]interface{}) error {
 	}
 
 	return json.Unmarshal(data, &output) // Convert to a map
-}
-
-// Concurrent Safe String keyed map
-type StringKeyedMap[T any] struct {
-	items map[string]T
-	lock  *sync.RWMutex
-}
-
-func NewStringKeyedMap[T any]() *StringKeyedMap[T] {
-	return &StringKeyedMap[T]{
-		items: make(map[string]T, 0),
-		lock:  &sync.RWMutex{},
-	}
-}
-
-func (sm *StringKeyedMap[T]) Store(key string, item T) {
-	sm.lock.Lock()
-	defer sm.lock.Unlock()
-	sm.items[key] = item
-}
-
-func (sm *StringKeyedMap[T]) Load(key string) (T, bool) {
-	sm.lock.RLock()
-	defer sm.lock.RUnlock()
-	ok, item := sm.items[key]
-	return ok, item
-}
-
-func (sm *StringKeyedMap[T]) Delete(key string) {
-	sm.lock.Lock()
-	defer sm.lock.Unlock()
-	delete(sm.items, key)
 }
